Update: Modularise Card component
