+++ test/components/adslotUi/AccordionComponentTest.jsx
@@ -4,7 +4,7 @@ import sinon from 'sinon';
 import { shallow, mount } from 'enzyme';
 import Accordion from 'components/adslotUi/AccordionComponent';
 import Panel from 'components/adslotUi/PanelComponent';
-import Card from 'components/alexandria/CardComponent';
+import Card from 'components/alexandria/Card/component';
 import PanelMocks from 'mocks/PanelMocks';
 
 describe('AccordionComponent', () => {
